requirements.txt
django
djangorestframework
matplotlib
yfinance
pandas
mysqlclient
celery
redis
FPDF
reportlab